Fix how map files work
The --map part isn't doing what we're thinking it does. It's not always a one-to-one mapping of rows.
For most of the data slinging that we do, it's more for translating a row of legacy ILS data unique ID to a Koha ID (borrowernumber, itemnumber, and biblionumber).
As an example scenario, I've got a file of reading history data (pretty much old_issues) with 2 million transactions.
They have two columns actor_id (borrower ID) and exem_id (item ID).
We store these IDs in a column in Koha so we could map them later; so actor_id is currently in sort2 and exem_id in stocknumber.
We would create a map for those columns (i.e. sort2, borrowernumber).
What we would like to see with the --map parameter is translate the actor_id or exem_id to its respective borrowernumber and itemnumber based on the map we have created.

I have a reading history data that I'd like to load in old_issues.  The data has actor_id, exem_id, date was checked out, date the item was due, and date the item has been returned.
actor_id is currently loaded into borrowers.sort2 and exem_id is loaded into items.stocknumber.
Typically, on migration, we create a map of borrowers.sort2,borrowers.borrowernumber (and the item equivalent).  We use the mapped file to link the legacy ILS data to.
Ideally the general masher would map the actor_id and exem_id to borrowernumber and itemnumber respectively based on the mapped file we created.

So, in other words, read <in file header> from the incoming csv file, then in the map file <mapping filename>,
find the row in the column <lookup mapping header> that matches the value,
pull the value from <output mapping header> and write it in the output file as <output header>

like `-m <in file header>:<mapping filename>:<lookup mapping header>:<output mapping header>:<output header>`